added tests for new webhook payload and old POST service payload
- refactored existing BitbucketHookReceiver class into 3 specialized components:
  - BitbucketHookReceiver: still responsible for wiring up to Jenkins
  - BitbucketPayloadProcessor: handles determinng type of payload and extracting data from JSON
  - BitbucketJobProbe: responsible for finding jobs to trigger matching the data extracted
