+++ pom.xml
@@ -63,6 +63,12 @@
             <artifactId>git</artifactId>
             <version>2.0.3</version>
         </dependency>
+        <dependency>
+            <groupId>org.mockito</groupId>
+            <artifactId>mockito-core</artifactId>
+            <version>2.0.22-beta</version>
+            <scope>test</scope>
+        </dependency>
     </dependencies>
 
 </project>
